requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
openai
weasyprint
jinja2
stripe
